Issue #3098898: Include completed drafts in new submission downloads
Previously, downloading new submissions would download submissions with
a submission ID larger than the largest one in the previous download.
This would exclude a submission which has an old sid because it was
saved as a draft but which has now been completed.

This commit makes downloading new submissions the same as range_date
except that the date range starts with the time of the last last
download.