Ensure #attributes for components is always an array
We have seen situations for current installs where #attributes is
stored as an empty string in the data base. This may be the result of
A bad upgrade path or the like. We do not know.

When upgrading to PHP 7.4 this results in errors “Cannot use string 
offset as an array” when rendering the affected components.

Updating each component through the admin interface fixes this but it
is unreasonable to require editors to do this for all webforms. 

To address this we ensure #attributes default to an empty array.

Another option might be an update hook going through all components.